Fixes bugs where the circle and triangle overlays and the watermark were not displayed correctly.
The circle and triangle overlays were previously handled by `CursorOverlay`, a separate top-level window that was not part of the video frame. This caused them to not be displayed correctly and not be included in recordings.

This commit merges the functionality of `CursorOverlay` into `VideoOverlay`, ensuring that the cursor highlights are drawn directly onto the video frame. This also resolves any potential display issues with the watermark.

The `CursorOverlay.py` file has been deleted as it is no longer needed.
